+++ backend/app.js
@@ -1,6 +1,7 @@
 global.mongoose = require("mongoose");
 const express = require("express");
 const user = require("./api/userService");
+const auth = require("./api/authService");
 const banlist = require("./api/banService");
 const post = require("./api/postService");
 const session = require("express-session");
@@ -21,6 +22,7 @@ app.use(
 
 const url = process.env.URL;
 user(app);
+auth(app);
 banlist(app);
 post(app);
